refactor: update FedaPay integration to EasySwitch payment gateway
EasySwitch.types : remove all FedaPay'specific types; EasySwitch.integrators.fedapay, EasySwitch.utils.__init__: refactoring of FedaPay'module.
